requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Werkzeug==2.0.1
passlib==1.7.4
email-validator==1.1.3
Flask-Cors==3.0.10
PyJWT==2.1.0
Pillow==8.3.1
python-dotenv==0.19.0 